% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runShinyGenotyping}
\alias{runShinyGenotyping}
\title{Launch ShinyGenotyping GUI}
\usage{
runShinyGenotyping()
}
\description{
This function run the GUI shiny interface to run ShinyGenotyping
}
\examples{
runShinyGenotyping()
}
